feat(api): 1. move ActivePermissionOperationsUtils from utils to account dir, 2 del unuseful blank line and add some notes
